+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Development
 
+### Added
+
+- `subtitleEnabled` and `subtitleLanguage` in outgoing payload 
 - `audioLanguage` in outgoing payload
 
 ## 1.7.5